Remove fixed tile size, turn script into function.
Turned the script of tile-add.js into a single function
within itself, labeled gridSet(num). Added validation
to the function to ensure that the number is < 100.

Need to add functionality to overwrite the existing container
with javascript upon input of a new number, and need to
write out the input form in html.